imports_app/requirements.txt
streamlit>=1.22.0
pandas>=1.5.0
geopandas>=0.12.0
matplotlib>=3.6.0
cartopy>=0.21.0
sdmx>=2.2.0
requests>=2.28.0
pillow>=9.4.0